Remove never ending sending page view
