+++ pillarbox-analytics/src/main/java/ch/srgssr/pillarbox/analytics/PageViewTracker.kt
@@ -4,32 +4,13 @@
  */
 package ch.srgssr.pillarbox.analytics
 
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.ProcessLifecycleOwner
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
-import kotlinx.coroutines.launch
-
 /**
  * Page view tracker send page views only if the last page view is different.
- * It will send again the last page view when application come back to foreground.
  *
  * @property pageViewAnalytics The [PageViewAnalytics] implementation to send page views.
  */
 class PageViewTracker(private val pageViewAnalytics: PageViewAnalytics) : PageViewAnalytics {
     private var lastPageView: PageView? = null
-
-    init {
-        val processLifecycleOwner = ProcessLifecycleOwner.get()
-        processLifecycleOwner.lifecycleScope.launch {
-            processLifecycleOwner.repeatOnLifecycle(Lifecycle.State.RESUMED) {
-                val lastPageView = clear()
-                lastPageView?.let {
-                    sendPageView(it)
-                }
-            }
-        }
-    }
 
     override fun sendPageView(pageView: PageView) {
         if (lastPageView != pageView) {
@@ -40,6 +21,7 @@ class PageViewTracker(private val pageViewAnalytics: PageViewAnalytics) : PageVi
 
     /**
      * Clear [lastPageView]
+     * @return the last page view if any.
      */
     fun clear(): PageView? {
         val output = lastPageView
